add override docs, move overrides to tokens file
