+++ src/index.ts
@@ -15,5 +15,6 @@ export {
   buildColors as UNSTABLE_buildColors,
   roleVariants as UNSTABLE_roleVariants,
   toCssCustomPropertySyntax as UNSTABLE_toCssCustomPropertySyntax,
+  Tokens as UNSTABLE_Tokens,
 } from './utilities/theme';
 /* eslint-enable @typescript-eslint/camelcase */
